git-svn-id: https://svn.l3s.uni-hannover.de/learnweb/trunk/InterWebJ@1376 422a2c9f-66af-4668-9993-60dc3b0871b8
Former-commit-id: 460f99c3ffafd7faa9597444d1c17228c71bbb27

+++ InterWebJ-Core/src/de/l3s/interwebj/query/QueryResult.java
@@ -15,7 +15,7 @@ public class QueryResult
 	private List<ResultItem> resultItems;
 	private long elapsedTime;
 	private long createdTime;
-	private int totalResultCount;
+	private long totalResultCount;
 	
 
 	public QueryResult(Query query)
@@ -65,7 +65,7 @@ public class QueryResult
 	}
 	
 
-	public int getTotalResultCount()
+	public long getTotalResultCount()
 	{
 		return totalResultCount;
 	}
@@ -84,7 +84,7 @@ public class QueryResult
 	}
 	
 
-	public void setTotalResultCount(int totalResultCount)
+	public void setTotalResultCount(long totalResultCount)
 	{
 		this.totalResultCount = totalResultCount;
 	}
